Moved modal into bootstrap controls folder and made it a core modal implementation
