+++ addon/components/bootstrap/modal/footer/-close-button.js
@@ -1,0 +1,7 @@
+import Ember from 'ember';
+import layout from '../../../../templates/components/bootstrap/modal/footer/-close-button';
+
+export default Ember.Component.extend({
+  tagName: '',
+  layout
+});
