+++ addon/components/bootstrap/modal/header/-close-button.js
@@ -1,0 +1,6 @@
+import Ember from 'ember';
+import layout from '../../../../templates/components/bootstrap/modal/header/-close-button';
+
+export default Ember.Component.extend({
+  layout
+});
