--- addon/components/modal/-body.js
@@ -1,17 +1,0 @@
-import Ember from 'ember';
-import layout from '../../templates/components/modal/-body';
-import { PropTypes } from 'ember-prop-types';
-import { BuilderForPropTypes, BuilderForPropDefaults } from 'ember-bootstrap-controls/utils/prop-definition-tools';
-
-export const propDefinitions = {};
-
-export default Ember.Component.extend({
-  layout,
-  propTypes: BuilderForPropTypes(propDefinitions),
-
-  classNames: ['modal-body'],
-
-  getDefaultProps() {
-    return BuilderForPropDefaults(propDefinitions)
-  },
-});
