--- addon/components/modal/footer/-accept-button.js
@@ -1,7 +1,0 @@
-import Ember from 'ember';
-import layout from '../../../templates/components/modal/footer/-accept-button';
-
-export default Ember.Component.extend({
-  tagName: '',
-  layout
-});
