--- addon/components/modal/footer/-close-button.js
@@ -1,7 +1,0 @@
-import Ember from 'ember';
-import layout from '../../../templates/components/modal/footer/-close-button';
-
-export default Ember.Component.extend({
-  tagName: '',
-  layout
-});
