--- addon/components/modal/header/-close-button.js
@@ -1,6 +1,0 @@
-import Ember from 'ember';
-import layout from '../../../templates/components/modal/header/-close-button';
-
-export default Ember.Component.extend({
-  layout
-});
