+++ addon/templates/components/bootstrap/modal/footer/-accept-button.hbs
