+++ addon/templates/components/bootstrap/modal/footer/-close-button.hbs
