+++ addon/templates/components/bootstrap/modal/header/-close-button.hbs
