+++ addon/templates/components/ember-bootstrap-controls-freestyle.hbs
@@ -18,8 +18,5 @@
   {{freestyle/bootstrap-inputs/-search section=section}}
   {{freestyle/bootstrap-inputs/-range section=section}}
   {{freestyle/bootstrap-inputs/-checkbox section=section}}
-{{/freestyle-section}}
-
-{{#freestyle-section name='Bootstrap Modal' as |section|}}
-  {{freestyle/modal/modal-container section=section}}
+  {{freestyle/bootstrap/modal/-modal section=section}}
 {{/freestyle-section}}
