--- addon/templates/components/modal/-footer.hbs
@@ -1,13 +1,0 @@
-{{#if hasBlock}}
-  {{yield
-    (hash
-      acceptAction=acceptAction
-      acceptButton=(component 'modal/footer/-accept-button' action=acceptAction)
-      closeAction=closeAction
-      closeButton=(component 'modal/footer/-close-button' action=closeAction)
-    )
-  }}
-{{else}}
-  {{modal/footer/-close-button action=closeAction}}
-  {{modal/footer/-accept-button action=acceptAction}}
-{{/if}}
