--- addon/templates/components/modal/-header.hbs
@@ -1,15 +1,0 @@
-{{#if hasBlock}}
-  {{yield
-    (hash
-      title=(component 'modal/header/-title' title=title)
-      closeAction=closeAction
-      closeButton=(component 'modal/header/-close-button' action=closeAction)
-      showCloseInHeader=showCloseInHeader
-    )
-  }}
-{{else}}
-  {{modal/header/-title title=title}}
-  {{#if showCloseInHeader}}
-    {{modal/header/-close-button action=closeAction}}
-  {{/if}}
-{{/if}}
