--- addon/templates/components/modal/modal-container.hbs
@@ -1,35 +1,0 @@
-<div class="modal-dialog{{if centered " modal-dialog-centered" ""}}" role="document" tabindex='-1'>
-  <div class="modal-content">
-    {{#if hasBlock}}
-      {{yield
-        (hash
-          isOpen=isOpen
-          toggle=(action 'toggleModal')
-          close=(action 'closeAction')
-          accept=acceptAction
-          header=(component 'modal/-header'
-            title=title
-            showCloseInHeader=showCloseInHeader
-            closeAction=(action 'closeAction')
-          )
-          body=(component 'modal/-body' content=body)
-          footer=(component 'modal/-footer'
-            acceptAction=acceptAction
-            closeAction=(action 'closeAction')
-          )
-        )
-      }}
-    {{else}}
-      {{modal/-header
-        title=title
-        showCloseInHeader=showCloseInHeader
-        closeAction=(action 'closeAction')
-      }}
-      {{modal/-body content=body}}
-      {{modal/-footer
-        acceptAction=acceptAction
-        closeAction=(action 'closeAction')
-      }}
-    {{/if}}
-  </div>
-</div>
