+++ app/components/bootstrap/modal/-footer.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-bootstrap-controls/components/bootstrap/modal/-footer';
