+++ app/components/bootstrap/modal/footer/-accept-button.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-bootstrap-controls/components/bootstrap/modal/footer/-accept-button';
