+++ app/components/bootstrap/modal/footer/-close-button.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-bootstrap-controls/components/bootstrap/modal/footer/-close-button';
