+++ app/components/bootstrap/modal/header/-close-button.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-bootstrap-controls/components/bootstrap/modal/header/-close-button';
