+++ app/components/freestyle/bootstrap/modal/-modal.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-bootstrap-controls/components/freestyle/bootstrap/modal/-modal';
