--- app/components/freestyle/modal/modal-container.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-bootstrap-controls/components/freestyle/modal/modal-container';
