--- app/components/modal/-body.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-bootstrap-controls/components/modal/-body';
