--- app/components/modal/-header.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-bootstrap-controls/components/modal/-header';
