--- app/components/modal/footer/-close-button.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-bootstrap-controls/components/modal/footer/-close-button';
