--- app/components/modal/modal-container.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-bootstrap-controls/components/modal/modal-container';
